requirements.txt
flask
openai
python-dotenv
requests
cohere